# public/swagger.yaml
basePath: /
consumes:
    - application/json
definitions: {}
info:
    title: restapi
    version: 1.0.0
    license:
        name: ISC
    contact: {}
paths:
    /addOrder/create:
        post:
            operationId: AddNewOrder
            produces:
                - application/json
            responses:
                '200':
                    description: 'Create Order'
                    schema:
                        additionalProperties: true
                '422':
                    description: 'Missing Field'
            consumes:
                - application/json
            description: 'Create a new order'
            tags:
                - Orders
            security: []
            parameters:
                -
                    in: body
                    name: res
                    required: true
                    schema:
                        type: object
    /addOrder/all:
        get:
            operationId: GetOrders
            produces:
                - application/json
            responses:
                '200':
                    description: 'All Orders'
                    schema:
                        additionalProperties: true
                '403':
                    description: Error
                '422':
                    description: 'Missing Field'
            description: 'Get all orders'
            tags:
                - Orders
            security: []
            parameters: []
    /addOrder/details:
        post:
            operationId: GetOrderDetails
            produces:
                - application/json
            responses:
                '200':
                    description: 'Orders Details'
                    schema:
                        additionalProperties: true
                '403':
                    description: Error
                '422':
                    description: 'Missing Field'
            consumes:
                - application/json
            description: 'Get details from order'
            tags:
                - Orders
            security: []
            parameters:
                -
                    in: body
                    name: res
                    required: true
                    schema:
                        type: object
    /products/all:
        get:
            operationId: GetProducts
            produces:
                - application/json
            responses:
                '200':
                    description: 'All data'
                    schema:
                        additionalProperties: true
                '403':
                    description: Error
            description: 'Get all Products'
            tags:
                - Product
            security: []
            parameters: []
    /products/create:
        post:
            operationId: Create
            produces:
                - application/json
            responses:
                '200':
                    description: 'Create Product'
                    schema:
                        additionalProperties: true
                '403':
                    description: Error
                '422':
                    description: 'Missing Field'
            consumes:
                - application/json
            description: 'Create a new product'
            tags:
                - Product
            security: []
            parameters:
                -
                    in: body
                    name: res
                    required: true
                    schema:
                        type: object
    /products/byType:
        post:
            operationId: ByType
            produces:
                - application/json
            responses:
                '200':
                    description: 'Product By Type'
                    schema:
                        additionalProperties: true
                '403':
                    description: 'Not Found'
                '422':
                    description: 'Missing Field'
            consumes:
                - application/json
            description: 'Show product by type'
            summary: 'A concise summary.'
            tags:
                - Product
            security: []
            parameters:
                -
                    in: body
                    name: res
                    required: true
                    schema:
                        type: object
    /productsType/create:
        post:
            operationId: Create
            produces:
                - application/json
            responses:
                '200':
                    description: 'Product Type created'
                    schema:
                        additionalProperties: true
                '401':
                    description: 'Not Authenticated'
                '403':
                    description: Error
                '422':
                    description: 'Missing Field'
            consumes:
                - application/json
            description: 'Create a type of products'
            summary: 'A concise summary.'
            tags:
                - 'Product Type'
            security: []
            parameters:
                -
                    in: body
                    name: res
                    required: true
                    schema:
                        type: object
    /productsType/all:
        get:
            operationId: GetAll
            produces:
                - application/json
            responses:
                '200':
                    description: 'Product Type created'
                    schema:
                        additionalProperties: true
                '403':
                    description: Error
            description: 'Get all types of products'
            summary: 'A concise summary.'
            tags:
                - 'Product Type'
            security: []
            parameters: []
    /user:
        get:
            operationId: GetUser
            produces:
                - application/json
            responses:
                '200':
                    description: 'User Data'
                    schema:
                        additionalProperties: true
                '403':
                    description: Error
                '422':
                    description: 'Missing Field'
            tags:
                - User
            security: []
            parameters: []
    /user/update:
        put:
            operationId: UpdateUser
            produces:
                - application/json
            responses:
                '200':
                    description: 'update User'
                    schema:
                        additionalProperties: true
                '403':
                    description: Error
                '422':
                    description: 'Missing Field'
            consumes:
                - application/json
            tags:
                - User
            security: []
            parameters:
                -
                    in: body
                    name: res
                    required: true
                    schema:
                        type: object
    /user/login:
        post:
            operationId: LogIn
            produces:
                - application/json
            responses:
                '200':
                    description: 'Session Created'
                    schema:
                        additionalProperties: true
                '403':
                    description: Error
                '422':
                    description: 'Missing Field'
            consumes:
                - application/json
            description: 'Login a user'
            tags:
                - User
            security: []
            parameters:
                -
                    in: body
                    name: res
                    required: true
                    schema:
                        type: object
    /user/signin:
        post:
            operationId: SignIn
            produces:
                - application/json
            responses:
                '200':
                    description: 'Account created successfully.'
                    schema:
                        additionalProperties: true
                '403':
                    description: Error
                '422':
                    description: 'Missing Field'
            consumes:
                - application/json
            description: 'Create a user'
            tags:
                - User
            security: []
            parameters:
                -
                    in: body
                    name: res
                    required: true
                    schema:
                        type: object
    /user/signOut:
        post:
            operationId: SignOut
            produces:
                - application/json
            responses:
                '200':
                    description: Success
                    schema:
                        additionalProperties: true
            consumes:
                - application/json
            description: 'Logout a user'
            tags:
                - User
            security: []
            parameters:
                -
                    in: body
                    name: res
                    required: true
                    schema:
                        type: object
    /user/addFavorites:
        post:
            operationId: AddFavorites
            produces:
                - application/json
            responses:
                '200':
                    description: 'Product added to favorites successfully.'
                    schema:
                        additionalProperties: true
                '403':
                    description: Error
                '422':
                    description: 'Missing Field'
            consumes:
                - application/json
            description: 'Add a product to fuser favorites'
            summary: 'A concise summary.'
            tags:
                - User
            security: []
            parameters:
                -
                    in: body
                    name: res
                    required: true
                    schema:
                        type: object
    /user/delFavorites:
        delete:
            operationId: DelFavorites
            produces:
                - application/json
            responses:
                '200':
                    description: 'Favorite Deleted'
                    schema:
                        additionalProperties: true
                '422':
                    description: 'Missing Field'
            consumes:
                - application/json
            description: 'Delete a product from user favorites'
            tags:
                - User
            security: []
            parameters:
                -
                    in: body
                    name: res
                    required: true
                    schema:
                        type: object
    /user/allFavorites:
        get:
            operationId: GetAllFavorites
            produces:
                - application/json
            responses:
                '200':
                    description: 'Product added to favorites successfully.'
                    schema:
                        additionalProperties: true
                '403':
                    description: Error
                '422':
                    description: 'Missing Field'
            description: 'Get all the favorites from a user'
            tags:
                - User
            security: []
            parameters: []
produces:
    - application/json
swagger: '2.0'
securityDefinitions:
    api_key:
        type: apiKey
        name: access_token
        in: query
    tsoa_auth:
        type: oauth2
        authorizationUrl: 'http://swagger.io/api/oauth/dialog'
        flow: implicit
        scopes:
            'write:pets': 'modify things'
            'read:pets': 'read things'